Fix MSVC compilation errors

+++ tests/src/generate.cpp
@@ -16,7 +16,7 @@ using namespace license;
 using namespace std;
 
 void generate_license(const string& fname, const vector<string>& other_args) {
-	int argc = 4+other_args.size();
+	int argc = (int)(4+other_args.size());
 	const char** argv = new const char*[argc + 1];
 	unsigned int i=0;
 	argv[i++] = "lic-generator";
